Abstract edits and raggedbottom added to class.
Abstract edits: The abstract is no longer bold and spans the full page width. This makes it a little more readable and separates it more clearly from the main body.

raggedbottom: Twocolumn uses flushbottom by default but the use of non floating elements that are separate from the text yields odd paragraph spacings because of this. Allowing unequal column lengths fixes this by pushing the whitespace to the bottom of a column rather than spreading it between paragraphs.